requirements.txt
asgiref >= 3.5.0
Django >= 3.2.12
django-environ >= 0.8.1
djangorestframework >= 3.13.1
Pillow >= 9.0.1
pytz >= 2022.1
sqlparse >= 0.4.2
typing_extensions >= 4.1.1